Read desc
The distance and index of objects inside of your NavMenu are now spoken as you navigate over them.
General improvements to the pathfinder to improve navigation. You will mostly notice this immediately on Offense, where you can now actually leave and go back to your ship without getting stuck.
Improved the logic used to add items to the "Nearby Unlabeled Objects" NavMenu list to filter out out-of-bounds objects.
Fixed an issue where the pathfinder would attempt to pathfind to objects out-of-bounds and falsly state that you reached the object when you did not.
Fixed an issue where certain objects in your NavMenu, mainly the "Enter" and "Exit" Factory options would be spoken as "EntranceTeleportA" and "B".